requirements.txt
Flask
Flask-Login
tensorflow
djangorestframework